+++ src/toTrigger/types.ts
@@ -6,4 +6,10 @@ export interface ToTriggerOptions {
   selector: string
   event?: FireEventNames
   check?: boolean
+  /** select all text in a textarea, input, contentEditable, etc */
+  selectAllText?: boolean
+  /** clears all the text in a textarea, input, contentEditable, etc */
+  clearAllText?: boolean
+  /** enters text in a textarea, input, contentEditable */
+  enterText?: string
 }