+++ CHANGELOG.md
@@ -14,3 +14,5 @@
  * removes puppeteer, jest, jest-puppeteer from hard dependencies
 
 # 0.0.3
+
+ * toTrigger: selectAllText, clearAllText, enterText